[TASK] DS exporter: Return generated statements
This changes the responsibilities: Until now, the statements were added
to a graph that was injected into the exporter object. Now, all code
using the exporter can use the generated statements in whatever fashion.

The format returned by the export method is the same as is used by the
Erfurt library graph classes.