[FEATURE] Add data mapping
The data mapping (i.e. exporting real records) is done with the column
mapper, because this process heavily relies on the statements generated
for the data structure export (mainly because we need it for the data
type).

The current status of the data export is rather unstable - there is
still a lot of possible types missing, and there is no complete export
for a record (which should be done by a dedicated mapper class, as for
the data structure export)