[BUGFIX] Single value relations are incorrectly mapped
The mapping result for a single value relation was a blank node with a
pointer to the node with uid 0 if the value was not set (for integer
fields, which perfectly make sense for single-value relations with only
one allowed table, e.g. the General Record Storage Page). This commit
changes the behaviour to use rdf:nil as value for unset fields.

Still to be done is the export for single-relation fields with a value
set - here the value could be set directly, without using a blank node.